Move entry algorithm into new folder
