+++ plugin/algorithms/__init__.py
@@ -1,0 +1,1 @@
+from .superimpose_by_entry import *
